Stamping out React warnings
