+++ client/src/pages/Profile/Profile.css
@@ -1,0 +1,8 @@
+body {
+ color: white;
+ text-align: center;
+}
+
+.sessionBtn{
+   float: center;
+}
