+++ client/src/routes.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import { Redirect, Route, Router, Switch } from 'react-router-dom';
+// import { Redirect, Route, Router, Switch } from 'react-router-dom';
+import { Route, Router, Switch } from 'react-router-dom';
 import App from './App';
 // import Home from './Home/Home';
-import Profile from './pages/Profile';
+import Profile from './pages/Profile/Profile';
 import Sessions from './pages/Sessions/Sessions';
 import Callback from './components/Callback/Callback';
 // import LoginPage from './pages/LoginPage/LoginPage'
