+++ package.json
@@ -26,7 +26,6 @@
     "if-env": "^1.0.4",
     "moment": "^2.23.0",
     "mongoose": "^5.3.16",
-    "react-materialize": "^2.6.0",
     "react-router-dom": "^4.3.1"
   }
 }
